[#2] Improved edge semantics
Edges are now simple objects in their own right, rather than just
appearing is entires in a map keyed by edge label.

Each edge is either a fixed-(string-)value edge, a regex edge, or an
integer edge. Labels are interpreted as:

* ":<anything>"" - fixed label <anything>, even if it would otherwise look
  special
* "*" - regex edge with regex ".*"
* "re: <regex>" - regex edge
* "#" integer edge

When a move is requested, edges are considered in the order they were
added to the machine, and the first matching edge is used. Results for
moves are cached so the same move in the future won't require an edge
scan.

V2 and V3 state machines have had their "*" edges tightened with regexes
as needed. A vendor-extension in the paths objectof the v3 walk test
model was previously misconstrued as a path object and caused the test
to fail. That test now passes, so things appear to be working, but more
extensive tests of different edge types should be added to the state
machine unit tests.

Also, javadocs need to be updated on the StateMachine class.