[#6] Added callback capability to JsonStateWalker
Also reworked terminology so that "walk" now refers to what the walker
does to the tree as a whole or to container nodes within it, while
"visit" is used for what happens to each node included in the walk.

The Disposition class for the result of a visit now includes an optional
Callback metnod instance, which will be invoked just prior to finishing
up the visit. This means that the walk method can do something while
visiting a container node and then undo it after all its children have
been visited.

We'll use that for managing a set of references already encountered in
this walk while visiting an ancestor node, in order to support reference
cycle detection.