[#6] Fix incorrect walk logic
We need to descend into child nodes of the current node, even if we're
not actually visiting the current node because we're in an anonymous
state. After all, anonymous states appear becuase they're intermediate
states in transits that lead to named nodes, so not descending in this
case could mean failing to visit some nodes corresponding to named states.